+++ R/reschola-package.R
@@ -4,7 +4,8 @@
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start
-#' @importFrom tibble tibble
 #' @importFrom lifecycle deprecate_soft
+#' @importFrom tibble tibble
+#' @importFrom utils packageVersion
 ## usethis namespace: end
 NULL